requirements.txt
accelerate==0.23.0
datasets==2.14.5
evaluate==0.4.1
einops==0.7.0
h5py==3.9.0
imageio==2.33.0
kornia==0.7.2 
kornia-rs==0.1.3
lightning==2.0.9
matplotlib==3.7.1
matplotlib-inline==0.1.6
monai==1.2.0
nibabel==5.1.0
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
protobuf==4.22.3
requests==2.28.1
safetensors==0.3.1
scikit-image==0.23.1
scikit-learn==1.2.2
scipy==1.13.0
termcolor==2.3.0
timm==0.6.13
torch-geometric==2.3.0
torch-summary==1.4.5
torchmetrics==0.11.4
torchsummary==1.5.1
tqdm==4.65.0
typing_extensions==4.8.0
wandb==0.15.12

